+++ packages/pulse_dev_flutter/test/pulse_test.dart
@@ -1,9 +1,12 @@
 import 'package:flutter_test/flutter_test.dart';
+import 'package:pulse_dev/pulse_dev.dart';
 import 'package:pulse_dev_flutter/pulse_dev_flutter.dart';
 
 import 'helpers/test_helpers.dart';
 
 void main() {
+  TestWidgetsFlutterBinding.ensureInitialized();
+
   setUp(() => Pulse.reset());
   tearDown(() => Pulse.reset());